Move runQueryStr to BazelUtil
