+++ multiversion/src/main/scala/multiversion/BazelUtil.scala
@@ -22,12 +22,31 @@ object BazelUtil {
       "package"
     )
 
-    bazel(app, bazelBin, command).map { out =>
+    bazelBytes(app, bazelBin, command).map { out =>
       app.env.workingDirectory.resolve(out.trim())
     }
   }
 
-  def bazel(
+  def runQueryStr(
+      app: Application,
+      bazelBin: Path,
+      queryExpression: String,
+      extraArgs: List[String]
+  ): Result[List[String]] = {
+    val command = List(
+      "query",
+      queryExpression,
+      "--notool_deps",
+      "--noimplicit_deps"
+    ) ++ extraArgs
+
+    bazelLines(app, bazelBin, command)
+  }
+
+  /**
+   * Run the Bazel command and capture the generated bytes via stderr.
+   */
+  def bazelBytes(
       app: Application,
       bazelBin: Path,
       command: List[String],
@@ -50,5 +69,24 @@ object BazelUtil {
     } else {
       pr0.asErrorResult(process.exitCode)
     }
+  }
+
+  /**
+   * Run the Bazel command and capture the generated lines via stdout.
+   */
+  def bazelLines(
+      app: Application,
+      bazelBin: Path,
+      command: List[String]
+  ): Result[List[String]] = {
+    val lines =
+      app
+        .process((bazelBin.toString :: command).map(x => (x: os.Shellable)): _*)
+        .call()
+        .out
+        .text()
+        .linesIterator
+        .toList
+    ValueResult(lines)
   }
 }
